move logic to distroHelper
