+++ integration-tests/src/test/resources/integration-test/openmrs-distro.properties
@@ -13,3 +13,4 @@ spa.frontendModules.@openmrs/esm-login-app=3.3.1
 spa.frontendModules.@openmrs/esm-patient-chart-app=3.1.0
 spa.apiUrl=notopenmrs
 spa.configUrls=foo
+content.hiv=1.0.0
