+++ maven-plugin/src/main/java/org/openmrs/maven/plugins/Setup.java
@@ -16,7 +16,6 @@ import org.openmrs.maven.plugins.model.Server;
 import org.openmrs.maven.plugins.model.Version;
 import org.openmrs.maven.plugins.utility.DBConnector;
 import org.openmrs.maven.plugins.utility.DistroHelper;
-import org.openmrs.maven.plugins.utility.PropertiesUtils;
 import org.openmrs.maven.plugins.utility.SDKConstants;
 import org.openmrs.maven.plugins.utility.ServerHelper;
 
@@ -277,7 +276,7 @@ public class Setup extends AbstractServerTask {
 			// `setServerVersionsFromDistroProperties`, and `server.setValuesFromDistroPropertiesModules`.
 			distroHelper.savePropertiesToServer(distroProperties, server);
 			setServerVersionsFromDistroProperties(server, distroProperties);
-			PropertiesUtils.parseContentProperties(distroProperties);
+			distroHelper.parseContentProperties(distroProperties);
 			moduleInstaller.installModulesForDistro(server, distroProperties, distroHelper);
 			setConfigFolder(server, distroProperties);
 			if (spaInstaller != null) {
