+++ sdk-commons/src/main/java/org/openmrs/maven/plugins/utility/DistroHelper.java
@@ -12,6 +12,8 @@ import org.openmrs.maven.plugins.model.PackageJson;
 import org.openmrs.maven.plugins.model.Server;
 import org.openmrs.maven.plugins.model.UpgradeDifferential;
 import org.openmrs.maven.plugins.model.Version;
+import org.semver4j.Semver;
+import org.semver4j.SemverException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.twdata.maven.mojoexecutor.MojoExecutor;
@@ -20,6 +22,7 @@ import org.twdata.maven.mojoexecutor.MojoExecutor.Element;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Enumeration;
 import java.util.List;
@@ -30,7 +33,6 @@ import java.util.zip.ZipFile;
 
 import static org.openmrs.maven.plugins.model.BaseSdkProperties.PROPERTY_DISTRO_ARTIFACT_ID;
 import static org.openmrs.maven.plugins.model.BaseSdkProperties.PROPERTY_DISTRO_GROUP_ID;
-import static org.openmrs.maven.plugins.utility.PropertiesUtils.processContentProperties;
 import static org.twdata.maven.mojoexecutor.MojoExecutor.artifactId;
 import static org.twdata.maven.mojoexecutor.MojoExecutor.configuration;
 import static org.twdata.maven.mojoexecutor.MojoExecutor.element;
@@ -296,102 +298,6 @@ public class DistroHelper {
 		}
 
 		return distroProperties;
-	}
-
-	/**
-	 * Downloads and processes content packages specified in the given distro properties.
-	 *
-	 * <p>This method filters out properties starting with a specific prefix (defined by {@code CONTENT_PREFIX})
-	 * from the {@code distroProperties} file, identifies the corresponding versions, and downloads the
-	 * associated ZIP files from the Maven repository. It then processes each downloaded ZIP file to locate
-	 * and parse a {@code content.properties} file, ensuring that the content package is valid and meets
-	 * the expected requirements.</p>
-	 *
-	 * <p>If a {@code groupId} is overridden for a particular content package, the method uses the overridden
-	 * value when fetching the package from Maven. The ZIP files are temporarily stored and processed to extract
-	 * the {@code content.properties} file, which is then validated and compared against the dependencies specified
-	 * in the {@code distro.properties} file.</p>
-	 *
-	 * @param contentPackageZipFile The directory where content package ZIP files will be temporarily stored.
-	 * @param distroProperties The {@code DistroProperties} object containing key-value pairs that specify
-	 *                         content packages and other properties needed to build a distribution.
-	 *
-	 * @throws MojoExecutionException If there is an error during the download or processing of the content packages,
-	 *                                such as missing or invalid {@code content.properties} files, or any IO issues.
-	 */
-	public void downloadContentPackages(File contentPackageZipFile, DistroProperties distroProperties)
-	        throws MojoExecutionException {
-		Properties contentProperties = new Properties();
-		
-		for (Object key : distroProperties.getAllKeys()) {
-			String keyOb = key.toString();
-			if (!keyOb.startsWith(CONTENT_PREFIX)) {
-				continue;
-			}
-			
-			String version = distroProperties.get(keyOb);
-			String zipFileName = keyOb.replace(CONTENT_PREFIX, "") + "-" + version + ".zip";
-			String artifactId = keyOb.replace(CONTENT_PREFIX, "");
-			String groupId = checkIfOverwritten(keyOb, distroProperties);
-			
-			Artifact artifact = new Artifact(artifactId, version, groupId, "zip");
-			File zipFile = downloadDistro(contentPackageZipFile, artifact, zipFileName);
-			
-			if (zipFile == null) {
-				log.warn("ZIP file not found for content package: {}", keyOb);
-				continue;
-			}
-
-			try (ZipFile zip = new ZipFile(contentPackageZipFile)) {
-				boolean foundContentProperties = false;
-				Enumeration<? extends ZipEntry> entries = zip.entries();
-				
-				while (entries.hasMoreElements()) {
-					ZipEntry zipEntry = entries.nextElement();
-					if (zipEntry.getName().equals(CONTENT_PROPERTIES)) {
-						foundContentProperties = true;
-						
-						try (InputStream inputStream = zip.getInputStream(zipEntry)) {
-							contentProperties.load(inputStream);
-							log.info("content.properties file found in {} and parsed successfully.",
-							    contentPackageZipFile.getName());
-							
-							if (contentProperties.getProperty("name") == null
-							        || contentProperties.getProperty("version") == null) {
-								throw new MojoExecutionException(
-								        "Content package name or version not specified in content.properties in "
-								                + contentPackageZipFile.getName());
-							}
-							
-							processContentProperties(contentProperties, distroProperties, contentPackageZipFile.getName());
-						}
-						break;
-					}
-				}
-				
-				if (!foundContentProperties) {
-					throw new MojoExecutionException(
-					        "No content.properties file found in ZIP file: " + contentPackageZipFile.getName());
-				}
-				
-			}
-			catch (IOException e) {
-				throw new MojoExecutionException("Error reading content.properties from ZIP file: "
-				        + contentPackageZipFile.getName() + ": " + e.getMessage(), e);
-			}
-		}
-	}
-
-	/**
-	 * Checks if the groupId for a given dependency is overwritten in the distro properties.
-	 *
-	 * @param dependencyKey    The key of the dependency.
-	 * @param distroProperties The distro properties file.
-	 * @return The groupId to use for this dependency.
-	 */
-	private static String checkIfOverwritten(String dependencyKey, DistroProperties distroProperties) {
-		String groupId = distroProperties.get(dependencyKey + ".groupId");
-		return groupId != null ? groupId : "org.openmrs.content";
 	}
 
 	public DistroProperties downloadDistroProperties(File serverPath, Server server, String fileType) throws MojoExecutionException {
@@ -670,6 +576,215 @@ public class DistroHelper {
 		server.setDistroGroupId(parentArtifact.getGroupId());
 		server.setVersion(parentArtifact.getVersion());
 		return resolveParentArtifact(parentArtifact, server.getServerDirectory(), distroProperties, appShellVersion);
+	}
+
+	/**
+	 * Parses and processes content properties from content packages defined in the given {@code DistroProperties} object.
+	 *
+	 * <p>This method creates a temporary directory to download and process content package ZIP files specified
+	 * in the {@code distroProperties} file. The method delegates the download and processing of content packages
+	 * to the {@code downloadContentPackages} method, ensuring that the content packages are correctly handled
+	 * and validated.</p>
+	 *
+	 * <p>After processing, the temporary directory used for storing the downloaded ZIP files is deleted,
+	 * even if an error occurs during processing. If the temporary directory cannot be deleted, a warning is logged.</p>
+	 *
+	 * @param distroProperties The {@code DistroProperties} object containing key-value pairs specifying
+	 *                         content packages and other properties needed to build a distribution.
+	 *
+	 * @throws MojoExecutionException If there is an error during the processing of content packages,
+	 *                                such as issues with creating the temporary directory, downloading
+	 *                                the content packages, or IO errors during file operations.
+	 */
+	public void parseContentProperties(DistroProperties distroProperties) throws MojoExecutionException {
+		File tempDirectory = null;
+		try {
+			tempDirectory = Files.createTempDirectory("content-packages").toFile();
+			downloadContentPackages(tempDirectory, distroProperties);
+
+		} catch (IOException e) {
+			throw new MojoExecutionException("Failed to process content packages", e);
+		} finally {
+			if (tempDirectory != null && tempDirectory.exists()) {
+				try {
+					FileUtils.deleteDirectory(tempDirectory);
+				} catch (IOException e) {
+					log.warn("Failed to delete temporary directory: {}", tempDirectory.getAbsolutePath(), e);
+				}
+			}
+		}
+	}
+
+	/**
+	 * Downloads and processes content packages specified in the given distro properties.
+	 *
+	 * <p>This method filters out properties starting with a specific prefix (defined by {@code CONTENT_PREFIX})
+	 * from the {@code distroProperties} file, identifies the corresponding versions, and downloads the
+	 * associated ZIP files from the Maven repository. It then processes each downloaded ZIP file to locate
+	 * and parse a {@code content.properties} file, ensuring that the content package is valid and meets
+	 * the expected requirements.</p>
+	 *
+	 * <p>If a {@code groupId} is overridden for a particular content package, the method uses the overridden
+	 * value when fetching the package from Maven. The ZIP files are temporarily stored and processed to extract
+	 * the {@code content.properties} file, which is then validated and compared against the dependencies specified
+	 * in the {@code distro.properties} file.</p>
+	 *
+	 * @param contentPackageZipFile The directory where content package ZIP files will be temporarily stored.
+	 * @param distroProperties The {@code DistroProperties} object containing key-value pairs that specify
+	 *                         content packages and other properties needed to build a distribution.
+	 *
+	 * @throws MojoExecutionException If there is an error during the download or processing of the content packages,
+	 *                                such as missing or invalid {@code content.properties} files, or any IO issues.
+	 */
+	public void downloadContentPackages(File contentPackageZipFile, DistroProperties distroProperties)
+			throws MojoExecutionException {
+		Properties contentProperties = new Properties();
+
+		for (Object key : distroProperties.getAllKeys()) {
+			String keyOb = key.toString();
+			if (!keyOb.startsWith(CONTENT_PREFIX)) {
+				continue;
+			}
+
+			String version = distroProperties.get(keyOb);
+			String zipFileName = keyOb.replace(CONTENT_PREFIX, "") + "-" + version + ".zip";
+			String artifactId = keyOb.replace(CONTENT_PREFIX, "");
+			String groupId = checkIfOverwritten(keyOb, distroProperties);
+
+			Artifact artifact = new Artifact(artifactId, version, groupId, "zip");
+			File zipFile = downloadDistro(contentPackageZipFile, artifact, zipFileName);
+
+			if (zipFile == null) {
+				log.warn("ZIP file not found for content package: {}", keyOb);
+				continue;
+			}
+
+			try (ZipFile zip = new ZipFile(zipFile)) {
+				boolean foundContentProperties = false;
+				Enumeration<? extends ZipEntry> entries = zip.entries();
+
+				while (entries.hasMoreElements()) {
+					ZipEntry zipEntry = entries.nextElement();
+					if (zipEntry.getName().equals(CONTENT_PROPERTIES)) {
+						foundContentProperties = true;
+
+						try (InputStream inputStream = zip.getInputStream(zipEntry)) {
+							contentProperties.load(inputStream);
+							log.info("content.properties file found in {} and parsed successfully.",
+									contentPackageZipFile.getName());
+
+							if (contentProperties.getProperty("name") == null
+									|| contentProperties.getProperty("version") == null) {
+								throw new MojoExecutionException(
+										"Content package name or version not specified in content.properties in "
+												+ contentPackageZipFile.getName());
+							}
+
+							processContentProperties(contentProperties, distroProperties, contentPackageZipFile.getName());
+						}
+						break;
+					}
+				}
+
+				if (!foundContentProperties) {
+					throw new MojoExecutionException(
+							"No content.properties file found in ZIP file: " + contentPackageZipFile.getName());
+				}
+
+			}
+			catch (IOException e) {
+				throw new MojoExecutionException("Error reading content.properties from ZIP file: "
+						+ contentPackageZipFile.getName() + ": " + e.getMessage(), e);
+			}
+		}
+	}
+
+	/**
+	 * Processes the {@code content.properties} file of a content package and validates the dependencies
+	 * against the {@code DistroProperties} provided. This method ensures that the dependencies defined
+	 * in the {@code content.properties} file are either present in the {@code distroProperties} file with
+	 * a version that matches the specified version range, or it finds the latest matching version if not
+	 * already specified in {@code distroProperties}.
+	 *
+	 * <p>The method iterates over each dependency listed in the {@code content.properties} file, focusing on
+	 * dependencies that start with specific prefixes such as {@code omod.}, {@code owa.}, {@code war},
+	 * {@code spa.frontendModule}, or {@code content.}. For each dependency, the method performs the following:</p>
+	 * <ul>
+	 *   <li>If the dependency is not present in {@code distroProperties}, it attempts to find the latest version
+	 *   matching the specified version range and adds it to {@code distroProperties}.</li>
+	 *   <li>If the dependency is present, it checks whether the version specified in {@code distroProperties}
+	 *   falls within the version range specified in {@code content.properties}. If it does not, an error is thrown.</li>
+	 * </ul>
+	 *
+	 * @param contentProperties The {@code Properties} object representing the {@code content.properties} file
+	 *                          of a content package.
+	 * @param distroProperties  The {@code DistroProperties} object containing key-value pairs specifying
+	 *                          the current distribution's dependencies and their versions.
+	 * @param zipFileName       The name of the ZIP file containing the {@code content.properties} file being processed.
+	 *                          Used in error messages to provide context.
+	 *
+	 * @throws MojoExecutionException If no matching version is found for a dependency not defined in
+	 *                                {@code distroProperties}, or if the version specified in {@code distroProperties}
+	 *                                does not match the version range in {@code content.properties}.
+	 */
+	protected void processContentProperties(Properties contentProperties, DistroProperties distroProperties, String zipFileName) throws MojoExecutionException {
+		for (String dependency : contentProperties.stringPropertyNames()) {
+			if (dependency.startsWith("omod.") || dependency.startsWith("owa.") || dependency.startsWith("war")
+					|| dependency.startsWith("spa.frontendModule") || dependency.startsWith("content.")) {
+				String versionRange = contentProperties.getProperty(dependency);
+				String distroVersion = distroProperties.get(dependency);
+
+				if (distroVersion == null) {
+					String latestVersion = findLatestMatchingVersion(dependency, versionRange);
+					if (latestVersion == null) {
+						throw new MojoExecutionException(
+								"No matching version found for dependency " + dependency + " in " + zipFileName);
+					}
+					distroProperties.add(dependency, latestVersion);
+				} else {
+					checkVersionInRange(dependency, versionRange, distroVersion, contentProperties.getProperty("name"));
+				}
+			}
+		}
+	}
+
+	/**
+	 * Checks if the groupId for a given dependency is overwritten in the distro properties.
+	 *
+	 * @param dependencyKey    The key of the dependency.
+	 * @param distroProperties The distro properties file.
+	 * @return The groupId to use for this dependency.
+	 */
+	private static String checkIfOverwritten(String dependencyKey, DistroProperties distroProperties) {
+		String groupId = distroProperties.get(dependencyKey + ".groupId");
+		return groupId != null ? groupId : "org.openmrs.content";
+	}
+
+	/**
+	 * Checks if the version from distro.properties satisfies the range specified in content.properties.
+	 * Throws an exception if there is a mismatch.
+	 *
+	 * @param contentDependencyKey The key of the content dependency.
+	 * @param contentDependencyVersionRange The version range specified in content.properties.
+	 * @param distroPropertyVersion The version specified in distro.properties.
+	 * @param contentPackageName The name of the content package.
+	 * @throws MojoExecutionException If the version does not fall within the specified range or if the
+	 *             range format is invalid.
+	 */
+	private static void checkVersionInRange(String contentDependencyKey, String contentDependencyVersionRange, String distroPropertyVersion, String contentPackageName) throws MojoExecutionException {
+		Semver semverVersion = new Semver(distroPropertyVersion);
+
+		try {
+			boolean inRange = semverVersion.satisfies(contentDependencyVersionRange.trim());
+			if (!inRange) {
+				throw new MojoExecutionException("Incompatible version for " + contentDependencyKey + " in content package "
+						+ contentPackageName + ". Specified range: " + contentDependencyVersionRange
+						+ ", found in distribution: " + distroPropertyVersion);
+			}
+		} catch (SemverException e) {
+			throw new MojoExecutionException("Invalid version range format for " + contentDependencyKey
+					+ " in content package " + contentPackageName + ": " + contentDependencyVersionRange, e);
+		}
 	}
 
 	public String findLatestMatchingVersion(String dependency, String versionRange) {
